Refactoring di entity, repository, dto, service e controller

+++ src/main/java/com/example/ordinaMii/DTO/Response/DishLightResponseDTO.java
@@ -3,12 +3,15 @@ package com.example.ordinaMii.DTO.Response;
 import com.example.ordinaMii.Entity.Enum.DishCategory;
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Builder
-public record DishLightResponseDTO(UUID id,
-                                   String name,
-                                   double price,
-                                   boolean available,
-                                   DishCategory category) {
+public record DishLightResponseDTO(
+        UUID id,
+        String name,
+        BigDecimal price,
+        boolean available,
+        DishCategory category
+) {
 }
