+++ src/main/java/com/example/ordinaMii/DTO/Response/DishResponseDTO.java
@@ -3,12 +3,16 @@ package com.example.ordinaMii.DTO.Response;
 import com.example.ordinaMii.Entity.Enum.DishCategory;
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.util.UUID;
+
 @Builder
-public record DishResponseDTO(UUID id,
-                              String name,
-                              String description,
-                              double price,
-                              boolean available,
-                              DishCategory category) {
+public record DishResponseDTO(
+        UUID id,
+        String name,
+        String description,
+        BigDecimal price,
+        boolean available,
+        DishCategory category
+) {
 }
