+++ src/main/java/com/example/ordinaMii/DTO/Response/OrderItemResponseDTO.java
@@ -2,13 +2,14 @@ package com.example.ordinaMii.DTO.Response;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Builder
 public record OrderItemResponseDTO(
         UUID id,
         int quantity,
-        double unitPrice,
+        BigDecimal unitPrice,
         DishLightResponseDTO dish
 ) {
 }
